Working on upload functionality right now

+++ server/src/controllers/user-account/vendor-account/stores/products/index.ts
@@ -7,6 +7,7 @@ import {
 } from "../../../../../app-schema/products.js";
 import db from "../../../../../db/index.js";
 import BadRequestError from "../../../../../errors/bad-request.js";
+import UnauthenticatedError from "../../../../../errors/unauthenticated.js";
 import {
   Insert,
   Update,
@@ -26,11 +27,14 @@ const createQuery = [
   async ({ reqData, params, userId: vendorId }) => {
     // this makes sure the Store exists before accessing the /user/stores/products endpoint
     const { storeId } = params;
-    const dbQuery = await db.query("select 1 from stores where store_id=$1", [
-      storeId,
-    ]);
+    const dbQuery = await db.query(
+      "select vendor_id from stores where store_id=$1",
+      [storeId]
+    );
     if (!dbQuery.rowCount)
       throw new BadRequestError("Store does not exist. Create a store");
+    if (dbQuery.rows[0].vendor_id !== vendorId)
+      throw new UnauthenticatedError("Cannot access store.");
     return await db.query(
       `${Insert("products", [
         ...Object.keys(reqData),
@@ -43,14 +47,17 @@ const createQuery = [
 ];
 
 const readAllQuery = [
-  async ({ query, limit, offset, params }) => {
+  async ({ query, limit, offset, params, userId: vendorId }) => {
     let { sort } = query;
     const { storeId } = params;
-    const dbQuery = await db.query("select 1 from stores where store_id=$1", [
-      storeId,
-    ]);
+    const dbQuery = await db.query(
+      "select vendor_id from stores where store_id=$1",
+      [storeId]
+    );
     if (!dbQuery.rowCount)
       throw new BadRequestError("Store does not exist. Create a store");
+    if (dbQuery.rows[0].vendor_id !== vendorId)
+      throw new UnauthenticatedError("Cannot access store.");
     // TODO: you may not need to store anything except the public_id of image in db
     let queryString = `select products.*, (select json_agg(media) from (select * from product_media where product_id=products.product_id) as media) as media from products where store_id=$1`;
     if (sort) {
@@ -64,13 +71,16 @@ const readAllQuery = [
 ];
 
 const readQuery = [
-  async ({ params }) => {
+  async ({ params, userId: vendorId }) => {
     let { storeId, productId } = params;
-    const dbQuery = await db.query("select 1 from stores where store_id=$1", [
-      storeId,
-    ]);
+    const dbQuery = await db.query(
+      "select vendor_id from stores where store_id=$1",
+      [storeId]
+    );
     if (!dbQuery.rowCount)
       throw new BadRequestError("Store does not exist. Create a store");
+    if (dbQuery.rows[0].vendor_id !== vendorId)
+      throw new UnauthenticatedError("Cannot access store.");
     return await db.query(
       `select products.*, (select json_agg(media) from (select * from product_media where product_id=$1) as media) as media from products where product_id=$1`,
       [productId]
@@ -79,13 +89,16 @@ const readQuery = [
 ];
 
 const updateQuery = [
-  async ({ params, reqData }) => {
+  async ({ params, reqData, userId: vendorId }) => {
     const { productId, storeId } = params;
-    const dbQuery = await db.query("select 1 from stores where store_id=$1", [
-      storeId,
-    ]);
+    const dbQuery = await db.query(
+      "select vendor_id from stores where store_id=$1",
+      [storeId]
+    );
     if (!dbQuery.rowCount)
       throw new BadRequestError("Store does not exist. Create a store");
+    if (dbQuery.rows[0].vendor_id !== vendorId)
+      throw new UnauthenticatedError("Cannot access store.");
     const updateCommand = Update(
       "products",
       "product_id",
@@ -99,15 +112,16 @@ const updateQuery = [
 ];
 
 const deleteQuery = [
-  async ({ params }) => {
+  async ({ params, userId: vendorId }) => {
     const { productId, storeId } = params;
     const dbQuery = await db.query(
-      "select store_id from stores where store_id=$1",
+      "select vendor_id from stores where store_id=$1",
       [storeId]
     );
     if (!dbQuery.rowCount)
       throw new BadRequestError("Store does not exist. Create a store");
-    assert(storeId === dbQuery.rows[0].store_id);
+    if (dbQuery.rows[0].vendor_id !== vendorId)
+      throw new UnauthenticatedError("Cannot access store.");
     return await db.query(
       `delete from products where product_id=$1 and store_id=$2`,
       [productId, storeId]
